Implement Assignment local service
1. Add to the AssignmentLocalServiceImpl:
Implement the addAssignment() in the class as follows:

2. Implement updateAssignment
Create an overload for the updateAssignment()

3. Implement the Finder Methods
Defining finders in service.xml automatically creates the corresponding methods in
the persistence classes, but we cannot access those directly from the controller layer and
have to implement façades in the service implementation class.

4. "Silence" the Generated Method
Sometimes it's practical to silence generated methods to ensure correct API usage. Override
and "silence" the generated addAssignment() and updateAssignment() method
signatures, which we replaced with our overrides before.