Implement Validation
1. Open the
com.liferay.training.gradebook.exception.AssignmentValidationExc
eption.java in the gradebook-api module.
and Implement the following new constructor and code to the class

2. Implement an Assignment Validator Interface
Create an interface com.liferay.training.gradebook.validator.AssignmentValidator

3. Implement an Assignment Validator Service Component
Create a class com.liferay.training.gradebook.util.validator.AssignmentValidatorImpl

4. Add a reference to the AssignmentValidator service to the end of the class:
com.liferay.training.gradebook.service.impl.AssignmentLocalServiceImpl
@Reference
AssignmentValidator _assignmentValidator;

5. Start serviceBuilder

6. Implement Feedback Messages Dispatching on the Controller Layer in GradebookPortlet.java

Validation is now implemented on the service layer. As we call the services on the
controller layer through the MVC commands in the gradebook-web module, we have to
pass the feedback messages from the service layer to the user interface there.
For transporting the messages to the user interface, we'll be using the SessionMessages
object for success messages and the SessionErrors object for the error messages.
You've probably noticed the default success message the platform sets when you add an
Assignment. We'll silence that because we'll be using our custom message.

7.  Upgrade classes:

Modify the doProcessAction() methods of the three MVC Action Command classes
in the gradebook-web module, calling the service to set the success and error messages for
the user interface:
Update the code of all of the following MVC Action Command classes as follows

AddAssignmentMVCActionCommand.java
EditAssignmentMVCActionCommand.java
DeleteAssignmentMVCActionCommand.java

8. Add an import for the SessionErrors class for showing the error message details:
-  Add an import in src/main/resources/META-INF/resources/init.jsp

-  After we add, update, or delete an Assignment successfully, we are redirected to the main
list view, implemented with the view.jsp

-  Add <liferay-ui> tags to src/main/resources/METAINF/resources/assignment/edit_assigment.jsp just after the init.jsp

- Add <liferay-ui> tags to src/main/resources/METAINF/resources/assignment/edit_assigment.jsp just after the init.jsp
include