Integrate with the Asset Framework
1. Add the Required Fields to the Assignment Entity
The Asset Framework requires a certain set of fields (columns) from all the entities. We are
currently missing the status fields:
status: used to determine an entity's status in the workflow
statusByUserId: status audit field
statusByUserName: status audit field
statusDate: status audit field
We'll also add a status finder for listing Assignments by their status.
-  Open the service.xml in the gradebook-service module.
- Add the status fields (see the highlighted fields). The final service.xml

2. Start ServiceBuilder

3. Implement Asset Resource Management in the Local Service
As with permissions, the Asset resource lifecycle has to be kept in sync with your entity:
whenever you create, update, or delete an Assignment entity, you have to take care of the
Asset resource:
- Open the class com.liferay.training.gradebook.service.impl.AssignmentLocalServiceImpl

-Start ServiceBuilder

4. Implement an Asset Renderer Factory for the Assignments
The service layer is now ready for the Asset Framework. Next, we'll create the
AssetRenderer and AssetRendererFactory components in the gradebook-web to take care of
displaying the assets in a standard way, for example, in the Asset Publisher portlet.
First, we'll add the required dependencies to the gradebook-web.

5. Create an AssetRendererFactory component to take care of providing the AssetRenderer

6. Implement an AssetRenderer for the Assignments

7. Implement the JSP files
At the final step, we'll implement the JSP files for abstract and full content Asset views. If
you take a look at the getJspPath() method in the AssetRenderer just created, you'll
see how the file path is built.

Implement the JSP files: abstract.jsp and full_content.jsp